Refactored names and fixed import statements in TDDCycles module
