+++ server/src/modules/TDDCycles/Application/getTDDCyclesUseCase.ts
@@ -1,6 +1,5 @@
 import { IDBCommitsRepository } from "../Domain/IDBCommitsRepository";
-import { CommitDataObject } from "../Domain/commitInterfaces";
-import { CommitDTO } from "../Domain/CommitDataObject";
+import { TDDCycleDataObject } from "../Domain/TDDCycleDataObject";
 import { IGithubRepository } from "../Domain/IGithubRepository";
 export class GetTDDCyclesUseCase {
   private dbCommitRepository: IDBCommitsRepository;
@@ -20,7 +19,7 @@ export class GetTDDCyclesUseCase {
       if ((await this.dbCommitRepository.repositoryExists(owner, repoName))) {
         commitsToSave = await this.dbCommitRepository.getCommitsNotSavedInDB(owner, repoName, commitsFromGithub);
       }
-      const commitsInfoForTDDCycles = await this.getCommitsInforForTDDCycle(owner, repoName, commitsToSave);
+      const commitsInfoForTDDCycles = await this.githubRepository.getCommitsInforForTDDCycle(owner, repoName, commitsToSave);
       await this.saveCommitsToDB(owner, repoName, commitsInfoForTDDCycles);
       const commits = await this.dbCommitRepository.getCommits(owner, repoName);
       return commits;
@@ -30,44 +29,10 @@ export class GetTDDCyclesUseCase {
     }
   }
 
-  async getCommitsInforForTDDCycle(
-    owner: string,
-    repoName: string,
-    commits: CommitDataObject[]
-  ) {
-    try {
-      const commitsFromSha = await Promise.all(
-        commits.map(commit => this.githubRepository.getCommitInfoForTDDCycle(owner, repoName, commit.sha))
-      );
-
-      const commitsData = commitsFromSha.map(({ html_url, stats, commit, sha, coveragePercentage }) => ({
-        html_url,
-        stats: {
-          total: stats.total,
-          additions: stats.additions,
-          deletions: stats.deletions,
-        },
-        commit: {
-          date: commit.author.date,
-          message: commit.message,
-          url: commit.url,
-          comment_count: commit.comment_count,
-        },
-        sha,
-        coverage: coveragePercentage,
-      }));
-
-      return commitsData;
-    } catch (error) {
-      console.error("Error getting commits from SHA:", error);
-      throw new Error("Error getting commits from SHA");
-    }
-  }
-
   async saveCommitsToDB(
     owner: string,
     repoName: string,
-    newCommits: CommitDTO[]
+    newCommits: TDDCycleDataObject[]
   ) {
     try {
         await Promise.all(
