+++ server/src/modules/TDDCycles/Domain/IDBCommitsRepository.ts
@@ -1,10 +1,10 @@
 import { Pool } from "pg";
-import { CommitDTO } from "./CommitDataObject";
-import { CommitDataObject } from "./commitInterfaces";
+import { TDDCycleDataObject } from "./TDDCycleDataObject";
+import { CommitDataObject } from "./CommitDataObject";
 
 export interface IDBCommitsRepository {
   pool: Pool;
-  saveCommit(owner: string, repoName: string, commit: CommitDTO): Promise<any>;
+  saveCommit(owner: string, repoName: string, commit: TDDCycleDataObject): Promise<any>;
   getCommits(owner: string, repoName: string): Promise<any>;
   commitExists(owner: string, repoName: string, sha: string): Promise<any>;
   repositoryExists(owner: string, repoName: string): Promise<boolean>;
