+++ server/src/modules/TDDCycles/Domain/TestResultDataObject.ts
@@ -1,0 +1,7 @@
+export interface JobDB{
+    id:number,
+    sha:string,
+    owner:string,
+    reponame:string,
+    conclusion:string;
+}
